This fixes textioHQ/frontend#2303
There are two parts of this fix:
First, this exposes a public EditorState.replace API to be consumed by
outside callers. We have a need to change the contentState without
triggering the undostack or related behavior. This function exposes the
proper way in which to do so.

Secondly, when going down the native codepath route, there is a chance
that outside callers can change the contentState during the onChange()
handler. If that happens, we need to cancel the native behavior and
render the code normally.