Fix jest forever
For some reason, unless we run tests in series tests are taking forever to run. Frazier and I spent about an hour on this before moving on. Our best guess is that one of the dependencies updated in a way that takes more memory than usual, because the failing tests were hitting circle's 4GB memory limit