Fix IE paste
In IE we paste by shifting focus between a trap contenteditable div (which collects what the browser thinks the paste should be) and the regular draft contenteditable. This shift of focus causes the cursor to be placed at the beginning of the document after pasting unless we do a forceSelection here. This should effectively be a noop.