Remove lookUpwardForInlineStyle
It doesn't make sense to use the inline style at the end of the previous block. That means that if you style the last word of a paragraph bold, select the next paragraph, and type to overwrite it, the inserted characters are bold instead of using the inline style of the overwritten text.