Switch to using selection in beforeinput
The DOM selection during the beforeinput event is the same as the composition range (before the character is inserted). That means that we don't have to recompute the composition range ourselves. That selection is ephemeral, is never rendered, and is reset immediately afterwards by a selectionchange, but we can parse it during the beforeinput handler to get the correct offsets to replace with the event data.