Adds support for copy and past in Android.
This largely hooks up the existing handlers for cut, copy, and paste.  Cut however is handled as a copy and delete so that it fits into the other beforeInput events.

Copy works without modification.

Paste on Android needs a forceSelection to place the cursor in the correct place, and the most reliable way to keep the keyboard open after a paste appears to be putting it into a setImmediate. This isn't fool proof, but it usually works.