Exit composition mode when idle
When the user has not modfied the dom for a while, and the caret is in a safe location, we switch out of composition mode.

This allows async updates such as text highlighting and so forth.  It does however increase the opportunity for the composition state to get out of sync with draft.