Trim injected spaces at start of paragraph
iOS usually does the right thing when you're typing.  If you swipe, it correctly guesses that the next word _should_ be preceded by a space.  This is NOT true though at the beginning of a paragraph.  Draft's event handling prevents iOS from knowing though whether something is a new paragraph or not.  So we can watch for this case and remove the first space.