Prevent composition reconciliation errors
Content that requires `restoreEditorDOM` no longer gets flagged as `nativelyRenderedContent`.  When exiting composition mode rendering in the timeout, we now force a rerender in a `setImmediate` to force draft to clean up content.  This does cause the keyboard to flicker if there was a mutation, though this only practically seems to happen when starting a new block.